Move schema and mirror functionality from app to controller
